[refs #00235] Add watch task PoC
We need to be able to continually poll our Sass for changes, so a watch
task is required. This small proof of concept gets it sorta working—they
key is not to redirect our output (in.scss > out.css), but to use the
standard node-sass CLI method (in.scss out.css).